+++ .gitignore
@@ -1,5 +1,5 @@
-coverage/
-node_modules/
-lib/
-test-report.xml
-.idea
+/coverage/
+/node_modules/
+/lib/
+/test-report.xml
+/.idea
